fix: moved js code to typescript for service
